--- a/nestedtensor/csrc/conv2d.cpp
+++ b/nestedtensor/csrc/conv2d.cpp
@@ -22,20 +22,41 @@ Tensor NestedTensor_conv2d(
     IntArrayRef dilation,
     int64_t groups) {
   Tensor input = input_;
+  auto self_opt_sizes = get_opt_sizes(input);
   if (is_nested_tensor_impl(input) && !is_nested_tensor_impl(weight)) {
     if (get_dim(input) == 4 && !bias && weight.size(2) == 1 && weight.size(3) == 1 &&
         stride[0] == 1 && stride[1] == 1 &&
         padding[0] == 0 && padding[1] == 0 &&
         dilation[0] == 1 && dilation[1] == 1 &&
-        groups == 1
+        groups == 1 &&
+        *self_opt_sizes[0] &&
+        *self_opt_sizes[1]
       ) {
       at::Tensor input_buffer;
-      std::cout << "get_is_contiguous(input): " << get_is_contiguous(input) << std::endl;
       if (get_is_contiguous(input) && input.dtype() == torch::kHalf) {
         std::cout << "HERE" << std::endl;
-        input = input.transpose(1, 3);
-        input = NestedTensor_contiguous(input);
+        // input = input.transpose(1, 3);
+        // input = NestedTensor_contiguous(input);
         input_buffer = get_buffer(input);
+        Tensor nt_sizes_ =
+            get_efficient_nested_size(input).sizes().to(torch::kInt32);
+        Tensor nt_sizes_1 = at::native::narrow(nt_sizes_, 1, 1, 1);
+        Tensor nt_sizes_2 = at::native::narrow(nt_sizes_, 1, 2, 1);
+        Tensor nt_sizes_all = nt_sizes_1 * nt_sizes_2;
+        int* nt_sizes_all_ptr = nt_sizes_all.data_ptr<int>();
+        std::vector<int> numbers;
+        numbers.reserve(1 + (nt_sizes_all.size(0) * *self_opt_sizes[1]));
+        numbers.push_back(0);
+        int64_t index = 1;
+        for (int64_t i = 0; i < nt_sizes_all.size(0); i++) {
+          for (int64_t j = 0; j < *self_opt_sizes[1]; j++) {
+            numbers.push_back(numbers[index - 1] + nt_sizes_all_ptr[i]);
+            index++;
+          }
+        }
+        at::Tensor numbers_t = torch::tensor(numbers).to(torch::kInt32);
+        std::cout << "numbers_t: " << numbers_t << std::endl;
+        Tensor nt_sizes = numbers_t.to(torch::kCUDA);
         input_buffer = input_buffer.reshape({-1, weight.size(1)});
       } else {
         input = input.transpose(1, 3);